Create README.md
A skippers application for managing offers on a website is a tool that will facilitate the management and presentation of sailing cruise offers to customers. The application will offer a number of functionalities that will help manage and present offers in an attractive way.

The first and basic functionality is to add and edit cruise offers. Skippers will be able to enter detailed information about the offer, such as the route of the cruise, date, length of the cruise, number of seats available on board, yacht equipment and price. In the future, they will also have the ability to add photos of yachts and places to be visited during the cruise.

Another important functionality is the management of reservations. Skippers can access a list of all bookings and edit them. They can also view the crew list.